Create README.md
It's a pain having to figure out what's going on in undocumented database procedures. Sometimes these can be hundreds or even thousands of lines long, using many tables, views and other stored procedures. This tool attempts to make it a little easier to get started with legacy code by parsing it and producing output that can then be drawn using the GraphViz tool.